Organização de telas em packages, recyclerView do fragment Music
