+++ app/src/main/AndroidManifest.xml
@@ -9,9 +9,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainScreen"></activity>
+        <activity android:name=".main.MainScreen"></activity>
         <activity
-            android:name=".ActSplash"
+            android:name=".main.ActSplash"
             android:theme="@style/SplashTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -19,10 +19,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".AccPlaceholder" />
-        <activity android:name=".NewAccountActivity" />
-        <activity android:name=".LoginActivity" />
-        <activity android:name=".MainActivity" />
+        <activity android:name=".main.AccPlaceholder" />
+        <activity android:name=".activity.NewAccountActivity" />
+        <activity android:name=".activity.LoginActivity" />
+        <activity android:name=".main.MainActivity" />
     </application>
 
 </manifest>
