+++ app/src/main/app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".MainActivity">
+        <activity android:name=".main.MainActivity">
             <intent-filter>
                 <action android:name="android.listenapp.splash_screen.MainActivity"/>
                 <category android:name="android.intent.category.DEFAULT" />
