+++ app/src/main/java/com/example/listenapp/activity/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.example.listenapp;
+package com.example.listenapp.activity;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -9,11 +9,15 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 
+import com.example.listenapp.main.CustomSnackbar;
+import com.example.listenapp.main.MainScreen;
+import com.example.listenapp.R;
 
-public class NewAccountActivity extends AppCompatActivity {
 
-    EditText user, password, passwordRepeat;
-    Button confirm, login;
+public class LoginActivity extends AppCompatActivity {
+
+    EditText user, password;
+    Button confirm, newAcc;
     Bundle accInfo;
     CoordinatorLayout layout;
 
@@ -25,13 +29,12 @@ public class NewAccountActivity extends AppCompatActivity {
         setClicks();
     }
     private void findViews() {
-        setContentView(R.layout.activity_new_account);
-        user = findViewById(R.id.input_user);
-        password = findViewById(R.id.input_pw);
-        passwordRepeat = findViewById(R.id.input_pwrepeat);
-        confirm = findViewById(R.id.confirm_Button);
-        login = findViewById(R.id.toLogin_Button);
-        layout = findViewById(R.id.coordinator);
+        setContentView(R.layout.activity_login);
+        user = findViewById(R.id.user_login_log);
+        password = findViewById(R.id.pw_login_log);
+        confirm = findViewById(R.id.confirm_log_Button);
+        newAcc = findViewById(R.id.toNewAcc_Button);
+        layout = findViewById(R.id.coordinator_login);
     }
 
     private void bundleSetup() {
@@ -40,7 +43,7 @@ public class NewAccountActivity extends AppCompatActivity {
     }
     private void setClicks() {
         confirm.setOnClickListener(goTO(MainScreen.class, accInfo));
-        login.setOnClickListener(goTO(LoginActivity.class));
+        newAcc.setOnClickListener(goTO(NewAccountActivity.class));
 
     }
     private View.OnClickListener goTO(final Class umaClasse) {
@@ -56,8 +59,7 @@ public class NewAccountActivity extends AppCompatActivity {
         return new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                if(user.getText().toString().isEmpty() || password.getText().toString().isEmpty() ||
-                        passwordRepeat.getText().toString().isEmpty()){
+                if(user.getText().toString().isEmpty() || password.getText().toString().isEmpty()){
                     new CustomSnackbar().makeSB(layout, getString(R.string.empty_fields_msg),15,15).show();
                 }
                 else {
