+++ app/src/main/java/com/example/listenapp/fragment/PlaylistsFragment.java
@@ -1,4 +1,4 @@
-package com.example.listenapp;
+package com.example.listenapp.fragment;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -9,15 +9,17 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class ProfileFragment extends Fragment {
+import com.example.listenapp.R;
 
-    public ProfileFragment() {
+public class PlaylistsFragment extends Fragment {
+
+    public PlaylistsFragment() {
     }
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-       return inflater.inflate(R.layout.fragment_profile, container, false);
+       return inflater.inflate(R.layout.fragment_playlists, container, false);
 
     }
     @Override
