+++ app/src/main/java/com/example/listenapp/fragment/ProfileFragment.java
@@ -1,28 +1,25 @@
-package com.example.listenapp;
+package com.example.listenapp.fragment;
 
 import android.content.Context;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import java.util.ArrayList;
+import com.example.listenapp.R;
 
-public class MusicFragment extends Fragment {
+public class ProfileFragment extends Fragment {
 
-    public MusicFragment() {
+    public ProfileFragment() {
     }
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-       return inflater.inflate(R.layout.fragment_music, container, false);
+       return inflater.inflate(R.layout.fragment_profile, container, false);
 
     }
     @Override
