+++ app/src/main/java/com/example/listenapp/main/ActSplash.java
@@ -1,4 +1,4 @@
-package com.example.listenapp;
+package com.example.listenapp.main;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,6 +6,8 @@ import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 
 public class ActSplash extends AppCompatActivity {
+
+
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
