+++ app/src/main/java/com/example/listenapp/main/CustomSnackbar.java
@@ -1,4 +1,4 @@
-package com.example.listenapp;
+package com.example.listenapp.main;
 
 import android.support.design.widget.CoordinatorLayout;
 import android.support.design.widget.Snackbar;
