+++ app/src/main/java/com/example/listenapp/main/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.listenapp;
+package com.example.listenapp.main;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -6,6 +6,10 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
+
+import com.example.listenapp.R;
+import com.example.listenapp.activity.LoginActivity;
+import com.example.listenapp.activity.NewAccountActivity;
 
 public class MainActivity extends AppCompatActivity {
