+++ app/src/main/res/layout/activity_acc_placeholder.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AccPlaceholder">
+    tools:context=".main.AccPlaceholder">
 
     <TextView
         android:id="@+id/user_name"
