+++ app/src/main/res/layout/activity_main_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainScreen"
+    tools:context=".main.MainScreen"
     android:orientation="vertical"
     android:background="@drawable/background">
