+++ app/src/main/res/layout/fragment_news.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NewsFragment"
+    tools:context=".fragment.NewsFragment"
     android:orientation="vertical">
 
     <TextView
